python/sites/monitor/requirements.txt
certifi==2023.5.7
charset-normalizer==3.2.0
idna==3.4
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.0.3
